Update the hostapd and wpa_supplicant log clean check point
The original one is to use the device reset but the reset is called when reaching to the end of the test case.
The log may not preserve. It moves to the stop_ap_handler and stop_sta_handler. If the configuration does not
exist, then it cleans the logs.